Add use of FuzzTest (#3)
Add some basic FuzzTest tests. Move path and authority structs inside the parts struct. The authority struct is now optional<>. I've left a mess in the form of the ComposeSplitEqual fuzz. This work shows that we need a means for compose() to validate the fields of the parts struct that it is fed.